backend/requirements.txt
# Django
Django>=5.0,<7.0
django-environ>=0.11.2

# Django REST Framework
djangorestframework>=3.15.0
django-filter>=24.0
django-cors-headers>=4.3.0

# Autenticación JWT
djangorestframework-simplejwt>=5.3.0

# Imágenes (requerido por ImageField)
Pillow>=10.0.0

# Cloudinary (almacenamiento de imágenes)
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0

# Base de datos
psycopg2-binary>=2.9.9

# Pasarela de pago
mercadopago>=2.2.0

# Email transaccional (Brevo / Sendinblue)
sib-api-v3-sdk>=7.6.0

# Producción
gunicorn>=22.0.0
